for /x/ index page, show page immediately, request a build and show progress
instead of showing nothing before a build finishes, fetch the code and
show the page.

the page itself contains javascript to make a SSE (server-sent events)
request to request a build.

we do max runtime.numcpu+1 builds at the same time (configurable). users
get updates on the waiting queue (how many in front of them), when the
build starts, and if it failed/succeeded. on failure, they see an error
message, and link to a log.  when the build succeeds, they are redirected
to the /z/ URL.