build with "go get" so we get proper module versioning information in the binary
cannot get that with go build. we want runtime/debug.ReadBuildInfo()
to tell us the right module version and hash, instead of "(devel)".

the complication is that we cannot seem to tell "go get" where to write
the resulting binary. there is no "-o" flag, and GOBIN isn't allowed
when cross compiling. the binary always goes to the $GOPATH/bin, and we
want to share GOPATH because of sumdb updates and checking.  for cross
compiling, the binary goes to $GOPATH/bin/$GOOS_$GOARCH, so we have to
detect that too. then we have to be careful with builds happening at
the same time: we cannot allow that if they would create their binary
in the same path.