fix bug where we don't properly return an error when a build failed
because of err variable shadowing. an error while handling an error. we
would return success for a build with a nil result, causing us to
dereference the nil later on. this would happen when saving the error
message of a build failure would fail itself. this again can happen with
a full disk.