add support building stripped binaries, without symbol table and debug information
makes binaries smaller.

this adds a 9th field to the record. it is optional. old software won't
understand these records. we don't have an explicit version in there (not
great). this will only give trouble if you use "gobuild get" with an old
gobuild client talking to a new gobuild server.

for issue #5