always resolve go module versions for incoming requests
so we will resolve commit hashes, and branch names. because that is what the go
proxy resolves. we'll redirect the user to the canonical version, e.g.
pseudoversion.

"gobuild get" now looks up the latest goversion and latest module version
through the gobuild sumdb, we resolve it for the client. and we print what we
resolved to by default.