fix resolving version of just a module without version info when a new toolchain was released
the latest go toolchain was used, but not actually downloaded. must have broken
this in the previous release... the function name "installedSDK()" was
misleading. we often only use the first return value, "allowedSDK", but it
isn't actually installed, we must ensure that separately. the new name is
"listSDK()" it doesn't imply this installed falsehood.