limit the number of concurrent command executions initiated by web requests
sometimes we get crawled (from many different IPs) with many concurrent
requests. we have a coordinator and a queue for doing the builds, so that
shouldn't bring us down. but we do various go command invocations before we
really start on a build. eg to fetch the go module, to list the main modules
(commands), to check if cgo is required. and that can still cause trouble with
many concurrent requests. so we now limit those to 3 concurrently. they
shouldn't last too long. web requests will just have to wait. if they abort the
connection, the waiting is aborted too.